+++ spec/LIN3S/WPFoundation/PostTypes/Fields/FieldsSpec.php
@@ -12,6 +12,8 @@
 namespace spec\LIN3S\WPFoundation\PostTypes\Fields;
 
 use LIN3S\WordPressPhpSpecBridge\ObjectBehavior;
+use LIN3S\WPFoundation\PostTypes\Fields\Components\FieldComponentInterface;
+use LIN3S\WPFoundation\PostTypes\Fields\PostTypeAndTemplateFieldConnector;
 
 /**
  * Spec of Fields class.
@@ -20,11 +22,6 @@ use LIN3S\WordPressPhpSpecBridge\ObjectBehavior;
  */
 class FieldsSpec extends ObjectBehavior
 {
-    function let()
-    {
-        $this->beAnInstanceOf('fixtures\LIN3S\WPFoundation\Fields');
-    }
-
     function it_extends_fields()
     {
         $this->shouldHaveType('LIN3S\WPFoundation\PostTypes\Fields\Fields');
@@ -42,6 +39,7 @@ class FieldsSpec extends ObjectBehavior
 
     function it_should_connector()
     {
+        $this->beAnInstanceOf('fixtures\LIN3S\WPFoundation\Fields');
         $this->connector()->shouldReturn([
             [
                 [
@@ -61,5 +59,21 @@ class FieldsSpec extends ObjectBehavior
     function it_should_be_remove_screen_attributes()
     {
         $this->removeScreenAttributes();
+    }
+
+    function it_should_initialize_fields_with_name_from_connector(
+        FieldComponentInterface $component,
+        PostTypeAndTemplateFieldConnector $connector
+    ) {
+        $connector->connector()->shouldBeCalled()->willReturn([]);
+        $connector->name()->shouldBeCalled()->willReturn('name');
+        $component->init('name', [])->shouldBeCalled();
+
+        $this->beConstructedWith([
+            $component
+        ], $connector);
+
+        $this->shouldHaveType('LIN3S\WPFoundation\PostTypes\Fields\Fields');
+
     }
 }